Important Files
These 3 files are fundamental to running the code. Button press contains code for when a button is connected and pressed. MidiscreenNew plays a midi file and displays the name on a OLED screen. MainMidiFile is a compilation of the two and eliminates the need for Midiscreen to be imported into Buttonpress.