Too long since last commit.
This is the rough, fully working python code. It uses user input through a rotary encoder to toggle through an array, allowing the user to select an song. The song then plays by sending its data through TTYMIDI to a Metro mini with code uploaded, allowing it to be an input. The song only plays when a button press is detected.